feat(domain): migrate FlowManager from orchestra to domain layer
- Copy orchestra/flow_dispatch.py to domain/flow_manager.py
- Convert orchestra/flow_dispatch.py to thin adapter shell
- Export FlowManager from domain.__init__.py
- Add tests for domain layer import
- Maintain backward compatibility via orchestra adapter

Ref: #1624
